Vivian Meazza:
I attach the latest version of the SpitfireIIa Model for updating CVS. It's
pretty much done now. I still have to animate the pilot, provide a fuel
gauge, and the reflector gun sight. These will take some time however, so I
thought people might like to play with it the way it is.

We still await proper engine/geared propeller code from Andy, but it's quite
fun the with legacy code.