Vivian Meazza:
I attach an update to the Spitfire model - it has not been working for some time apparently.

The update includes the necessary files to correct the bug, and some improved eye candy. It also includes a model of the Seafire LIIIc. This is not a re-liveried Spitfire, but a model in it's own right, but it does share about 90% of the Spitfire's files. More can be read in the Pilot's notes enclosed in the updated files.